refactor: move Menu classes into separate files
